admin eliminan userio en proceso

+++ src/main/java/Inmobilaria/GyL/service/impl/PropertyService.java
@@ -1,9 +1,6 @@
 package Inmobilaria.GyL.service.impl;
 
-import Inmobilaria.GyL.entity.Appointment;
-import Inmobilaria.GyL.entity.ImageProperty;
-import Inmobilaria.GyL.entity.Property;
-import Inmobilaria.GyL.entity.User;
+import Inmobilaria.GyL.entity.*;
 import Inmobilaria.GyL.enums.PropertyStatus;
 import Inmobilaria.GyL.enums.PropertyType;
 import Inmobilaria.GyL.repository.PropertyRepository;
@@ -13,7 +10,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.transaction.Transactional;
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
@@ -61,9 +57,11 @@ public class PropertyService implements IPropertyService {
     }
 
     public void deleteProperty(Long id) {
-        Property property = pr.findById(id).get();
-        for (ImageProperty img : property.getImages()){
-            ips.deleteById(img.getId());
+        Optional<Property> property = pr.findById(id);
+        if(property.isPresent()) {
+            for (ImageProperty img : property.get().getImages()) {
+                ips.deleteById(img.getId());
+            }
         }
 
         pr.deleteById(id);
